ViewModels separated for Testing
