+++ src/Messenger.sln
@@ -23,7 +23,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessengerScreenshare", "Mes
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessengerCloud", "MessengerCloud\MessengerCloud.csproj", "{D677997A-8B4D-44A9-BC1A-4CDE8CBF596A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MessengerTestUI", "MessengerTestUI\MessengerTestUI.csproj", "{3A447E60-7703-4824-9BE9-BE111EACD785}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessengerTestUI", "MessengerTestUI\MessengerTestUI.csproj", "{3A447E60-7703-4824-9BE9-BE111EACD785}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MessengerViewModels", "MessengerViewModels\MessengerViewModels.csproj", "{A847A1AC-FE7A-4D3E-9F00-9F2AC85C0BC3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -67,6 +69,10 @@ Global
 		{3A447E60-7703-4824-9BE9-BE111EACD785}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3A447E60-7703-4824-9BE9-BE111EACD785}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3A447E60-7703-4824-9BE9-BE111EACD785}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A847A1AC-FE7A-4D3E-9F00-9F2AC85C0BC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A847A1AC-FE7A-4D3E-9F00-9F2AC85C0BC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A847A1AC-FE7A-4D3E-9F00-9F2AC85C0BC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A847A1AC-FE7A-4D3E-9F00-9F2AC85C0BC3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
