+++ src/MessengerViewModels/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace MessengerViewModels
+{
+    public class Class1
+    {
+
+    }
+}
