+++ src/MessengerViewModels/Commands/NavigateHomeCommand.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MessengerApp.Stores;
-using MessengerApp.ViewModels;
+using MessengerViewModels.Stores;
+using MessengerViewModels.ViewModels;
 using MessengerDashboard;
 
-namespace MessengerApp.Commands
+namespace MessengerViewModels.Commands
 {
-    internal class NavigateHomeCommand : CommandBase
+    public class NavigateHomeCommand : CommandBase
     {
         private readonly NavigationStore _navigationStore;
