+++ src/MessengerViewModels/Commands/NavigateMeetCommand.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using MessengerApp.Stores;
-using MessengerApp.ViewModels;
+using MessengerViewModels.Stores;
+using MessengerViewModels.ViewModels;
 using MessengerDashboard;
 
-namespace MessengerApp.Commands
+namespace MessengerViewModels.Commands
 {
-    internal class NavigateMeetCommand : CommandBase
+    public class NavigateMeetCommand : CommandBase
     {
         private readonly NavigationStore _navigationStore;
