+++ src/MessengerViewModels/MessengerViewModels.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0-windows</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MessengerCloud\MessengerCloud.csproj" />
+    <ProjectReference Include="..\MessengerContent\MessengerContent.csproj" />
+    <ProjectReference Include="..\MessengerDashboard\MessengerDashboard.csproj" />
+    <ProjectReference Include="..\MessengerNetworking\MessengerNetworking.csproj" />
+    <ProjectReference Include="..\MessengerScreenshare\MessengerScreenshare.csproj" />
+    <ProjectReference Include="..\MessengerWhiteboard\MessengerWhiteboard.csproj" />
+  </ItemGroup>
+
+</Project>
