+++ src/MessengerViewModels/Services/INavigationService.cs
@@ -1,0 +1,9 @@
+﻿using MessengerViewModels.ViewModels;
+
+namespace MessengerViewModels.Services
+{
+    public interface INavigationService
+    {
+        void Navigate();
+    }
+}
