+++ src/MessengerViewModels/ViewModels/MeetViewModel.cs
@@ -1,5 +1,5 @@
-﻿using MessengerApp.Stores;
-using MessengerApp.Commands;
+﻿using MessengerViewModels.Stores;
+using MessengerViewModels.Commands;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,9 +7,9 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
 
-namespace MessengerApp.ViewModels
+namespace MessengerViewModels.ViewModels
 {
-    internal class MeetViewModel : ViewModelBase
+    public class MeetViewModel : ViewModelBase
     {
         public string VMString => "Meet";
         public ICommand NavigateHomeCommand { get; }
